+++ bindings/python/CHANGELOG.md
@@ -6,6 +6,9 @@
 # Changes in Version 1.13.0 (2026/XX/XX)
 
 - Add Polars Extension Type support for the BSON Binary, Code, ObjectId, and Decimal128 types.
+- Add ability to use threads or multiprocesses in `find_arrow_all`, `find_pandas_all`,
+  `find_numpy_all`, and `find_polars_all`
+
 
 # Changes in Version 1.12.0 (2026/01/26)